Fix a few other enchantments not being applied correctly

+++ src/main/java/ckathode/weaponmod/item/RangedComponent.java
@@ -220,7 +220,7 @@ public abstract class RangedComponent extends AbstractWeaponComponent {
         }
         int damage = EnchantmentHelper.getEnchantmentLevel(Enchantments.POWER, itemstack);
         if (damage > 0) {
-            entity.setExtraDamage((float) damage);
+            entity.setExtraDamage(damage);
         }
         int knockback = EnchantmentHelper.getEnchantmentLevel(Enchantments.PUNCH, itemstack);
         if (knockback > 0) {
